# Do not generate logs for Velox user errors