Summary:
user errors are expected to surface to users or be
suppressed by try.
no need to log them per row, this can be expensive.

after:
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast_int##try_invalid                                        1.80s   554.97m
cast_int##try_valid                                         2.36ms    423.55
cast_int##valid                                             2.37ms    422.34
----------------------------------------------------------------------------
```
before:
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast_int##try_invalid                                        3.09s   324.03m
cast_int##try_valid                                         2.38ms    420.10
cast_int##valid                                             2.42ms    412.64
----------------------------------------------------------------------------
```

Differential Revision: D47340682